#6: replaced all floating-point numbers with BigDecimal

+++ src/main/java/com/example/springweb/controllers/product/ProductDto.java
@@ -1,13 +1,11 @@
 package com.example.springweb.controllers.product;
 
-import jakarta.persistence.Column;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
+
+import java.math.BigDecimal;
 
 @Data
 @NoArgsConstructor
@@ -17,6 +15,6 @@ public class ProductDto {
     private Integer id;
     private String name;
     private String description;
-    private double price;
+    private BigDecimal price;
     private String imageName;
 }
